Separate Web modules related to Federation
