+++ lib/federation/controllers/activity_pub_controller.ex
@@ -3,10 +3,11 @@
 # SPDX-License-Identifier: AGPL-3.0-only
 # Upstream: https://git.pleroma.social/pleroma/pleroma/blob/develop/lib/pleroma/web/activity_pub/activity_pub_controller.ex
 
-defmodule MobilizonWeb.ActivityPubController do
-  use MobilizonWeb, :controller
+defmodule Mobilizon.Federation.ActivityPubController do
+  use Mobilizon.Federation, :controller
 
-  alias Mobilizon.{Actors, Actors.Actor, Config}
+  alias Mobilizon.{Actors, Config}
+  alias Mobilizon.Actors.Actor
 
   alias Mobilizon.Federation.ActivityPub
   alias Mobilizon.Federation.ActivityPub.Federator
@@ -18,7 +19,7 @@ defmodule MobilizonWeb.ActivityPubController do
 
   action_fallback(:errors)
 
-  plug(MobilizonWeb.Plugs.Federating when action in [:inbox, :relay])
+  plug(Mobilizon.Federation.Plugs.Federating when action in [:inbox, :relay])
   plug(:relay_active? when action in [:relay])
 
   def relay_active?(conn, _) do
