+++ test/federation/plugs/federating_plug_test.exs
@@ -3,15 +3,15 @@
 # Copyright © 2017-2019 Pleroma Authors <https://pleroma.social/>
 # SPDX-License-Identifier: AGPL-3.0-only
 
-defmodule MobilizonWeb.Plug.FederatingTest do
+defmodule Mobilizon.Federation.Plug.FederatingTest do
   use MobilizonWeb.ConnCase
+
+  alias Mobilizon.Federation.Plugs.Federating
 
   test "returns and halt the conn when federating is disabled" do
     Mobilizon.Config.put([:instance, :federating], false)
 
-    conn =
-      build_conn()
-      |> MobilizonWeb.Plugs.Federating.call(%{})
+    conn = Federating.call(build_conn(), %{})
 
     assert conn.status == 404
     assert conn.halted
@@ -20,9 +20,7 @@ defmodule MobilizonWeb.Plug.FederatingTest do
   test "does nothing when federating is enabled" do
     Mobilizon.Config.put([:instance, :federating], true)
 
-    conn =
-      build_conn()
-      |> MobilizonWeb.Plugs.Federating.call(%{})
+    conn = Federating.call(build_conn(), %{})
 
     refute conn.status
     refute conn.halted
