+++ test/mobilizon_web/resolvers/event_resolver_test.exs
@@ -2,10 +2,13 @@ defmodule MobilizonWeb.Resolvers.EventResolverTest do
   use MobilizonWeb.ConnCase
   use Bamboo.Test
   use Oban.Testing, repo: Mobilizon.Storage.Repo
-  alias Mobilizon.Events
-  alias MobilizonWeb.{AbsintheHelpers, Email}
-  alias Mobilizon.Service.Workers.BuildSearchWorker
+
   import Mobilizon.Factory
+
+  alias Mobilizon.Events
+  alias Mobilizon.Service.Workers
+
+  alias MobilizonWeb.{AbsintheHelpers, Email}
 
   @event %{
     description: "some body",
@@ -155,7 +158,7 @@ defmodule MobilizonWeb.Resolvers.EventResolverTest do
 
       assert json_response(res, 200)["data"]["createEvent"]["title"] == "come to my event"
       {id, ""} = json_response(res, 200)["data"]["createEvent"]["id"] |> Integer.parse()
-      assert_enqueued(worker: BuildSearchWorker, args: %{event_id: id, op: :insert_search_event})
+      assert_enqueued(worker: Workers.BuildSearch, args: %{event_id: id, op: :insert_search_event})
     end
 
     test "create_event/3 creates an event and escapes title and description", %{
@@ -201,7 +204,7 @@ defmodule MobilizonWeb.Resolvers.EventResolverTest do
                "<b>My description</b> <img src=\"http://placekitten.com/g/200/300\" />"
 
       {id, ""} = res["data"]["createEvent"]["id"] |> Integer.parse()
-      assert_enqueued(worker: BuildSearchWorker, args: %{event_id: id, op: :insert_search_event})
+      assert_enqueued(worker: Workers.BuildSearch, args: %{event_id: id, op: :insert_search_event})
     end
 
     test "create_event/3 creates an event as a draft", %{conn: conn, actor: actor, user: user} do
@@ -238,7 +241,7 @@ defmodule MobilizonWeb.Resolvers.EventResolverTest do
       {event_id_int, ""} = Integer.parse(event_id)
 
       refute_enqueued(
-        worker: BuildSearchWorker,
+        worker: Workers.BuildSearch,
         args: %{event_id: event_id_int, op: :insert_search_event}
       )
 
@@ -371,7 +374,7 @@ defmodule MobilizonWeb.Resolvers.EventResolverTest do
       {event_id_int, ""} = Integer.parse(event["id"])
 
       assert_enqueued(
-        worker: BuildSearchWorker,
+        worker: Workers.BuildSearch,
         args: %{event_id: event_id_int, op: :insert_search_event}
       )
     end
@@ -827,7 +830,7 @@ defmodule MobilizonWeb.Resolvers.EventResolverTest do
       {event_id_int, ""} = Integer.parse(event_res["id"])
 
       assert_enqueued(
-        worker: BuildSearchWorker,
+        worker: Workers.BuildSearch,
         args: %{event_id: event_id_int, op: :update_search_event}
       )
